修复删除功能

+++ dbapp/views/undergrade.py
@@ -120,12 +120,12 @@ def undergrade_search(request):
 @csrf_exempt
 def undergrade_deleteAll(request):
     data=request.POST
-    print(data)
     deletelist = data.getlist('vals')
-    print(deletelist)
     for i in range(len(deletelist)):
-        print(deletelist[i])
-        models.Undergraduate.objects.filter(id=int(deletelist[i])).delete()
+        if deletelist[i] != "":
+            models.Undergraduate.objects.filter(id=int(deletelist[i])).delete()
+        else:
+            continue
     return JsonResponse({'status':True})
 #本科生就业信息统计
 def undergrade_chart(request):
